Split main.rs into modules, make things easier to test
Now, everything fits on a screen, and the interfaces are a bit nicer -
next thing, I should probably add tests for the ssh_config parser.
(And also maybe figure out how to OAOO the config generation.)
