--- src/ssh_config.rustpeg
@@ -1,1 +1,0 @@
-// ssh_config rules here
